Move option package into cmd
